add: insert price feature

+++ sql/03-prices.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS prices (
     id INT PRIMARY KEY generated always as identity,
     ticker_id INT REFERENCES tickers(id),
-    price MONEY,
+    price FLOAT,
     price_date DATE NOT NULL
 );
